Rename `overview_screen` to `worker_overview_screen`
