+++ crates/hyperqueue/src/dashboard/ui/screens/root_screen.rs
@@ -2,7 +2,7 @@ use crate::dashboard::data::DashboardData;
 use crate::dashboard::ui::screen::Screen;
 use crate::dashboard::ui::screens::autoalloc_screen::AutoAllocScreen;
 use crate::dashboard::ui::screens::job_screen::JobScreen;
-use crate::dashboard::ui::screens::overview_screen::WorkerOverviewScreen;
+use crate::dashboard::ui::screens::worker_overview_screen::WorkerOverviewScreen;
 use crate::dashboard::ui::terminal::{DashboardFrame, DashboardTerminal};
 use crate::dashboard::{DEFAULT_LIVE_DURATION, TIMELINE_MOVE_OFFSET};
 use chrono::Local;
