+++ crates/hyperqueue/src/dashboard/ui/screens/worker_overview_screen/mod.rs
@@ -1,7 +1,7 @@
 use crate::dashboard::data::DashboardData;
 use crate::dashboard::ui::screen::Screen;
-use crate::dashboard::ui::screens::overview_screen::overview::fragment::ClusterOverviewFragment;
-use crate::dashboard::ui::screens::overview_screen::worker::fragment::WorkerOverviewFragment;
+use crate::dashboard::ui::screens::worker_overview_screen::overview::fragment::ClusterOverviewFragment;
+use crate::dashboard::ui::screens::worker_overview_screen::worker::fragment::WorkerOverviewFragment;
 use crate::dashboard::ui::terminal::DashboardFrame;
 use crossterm::event::{KeyCode, KeyEvent};
 use ratatui::layout::Rect;
