+++ crates/hyperqueue/src/dashboard/ui/screens/worker_overview_screen/worker/worker_utilization_chart.rs
